+++ examples/3qubit_gates.jl
@@ -3,7 +3,6 @@ function RX_on_q3()
     println(">>>>> RX Gate on third qubit using U3Gate <<<<<")
  
     return Dict{String, Any}(
-    
     "num_qubits" => 3, 
     "maximum_depth" => 3,
     "elementary_gates" => ["U3_3", "Identity"], 
@@ -42,7 +41,6 @@ function toffoli()
     println(">>>>> Toffoli gate <<<<<")
  
     return Dict{String, Any}(
-    
     "num_qubits" => 3, 
     "maximum_depth" => 15,
     "elementary_gates" => ["T_1", "T_2", "T_3", "H_3", "CNot_1_2", "CNot_1_3", "CNot_2_3", "Tdagger_1", "Tdagger_2", "Tdagger_3", "Identity"], 
@@ -61,7 +59,6 @@ function toffoli_using_kronecker()
     println(">>>>> Toffoli gate using Kronecker <<<<<")
  
     return Dict{String, Any}(
-    
     "num_qubits" => 3,
     "maximum_depth" => 12,
     "elementary_gates" => ["T_3", "H_3", "CNot_1_2", "CNot_1_3", "CNot_2_3", "Tdagger_3", "I_1xT_2xT_3", "CNot_1_2xH_3", "T_1xTdagger_2xI_3", "Identity"], 
@@ -79,7 +76,6 @@ function toffoli_using_2qubit_gates()
     println(">>>>> Toffoli gate with controlled gates <<<<<")
 
     return Dict{String, Any}(
-    
     "num_qubits" => 3,
     "maximum_depth" => 5,
     "elementary_gates" => ["CV_1_3", "CV_2_3", "CV_1_2", "CVdagger_1_3", "CVdagger_2_3", "CVdagger_1_2", "CNot_2_1", "CNot_1_2", "Identity"],
@@ -92,12 +88,11 @@ end
 function CNot_1_3()
 
     return Dict{String, Any}(
-
     "num_qubits" => 3,
     "maximum_depth" => 8,
     # "elementary_gates" => ["CNot_1_2", "CNot_2_3", "Identity"], 
     "elementary_gates" => ["H_1", "H_3", "H_2", "CNot_2_1", "CNot_3_2", "Identity"],
-    "target_gate" => QCOpt.get_full_sized_gate("CNot_1_3", 3),
+    "target_gate" => QCOpt.get_unitary("CNot_1_3", 3),
     "objective" => "minimize_depth",
     )
 end
@@ -132,17 +127,16 @@ function toffoli_left()
     # Reference: https://arxiv.org/pdf/0803.2316.pdf 
 
     function target_gate()
-        H_3 = QCOpt.get_full_sized_gate("H_3", 3)
-        Tdagger_3 = QCOpt.get_full_sized_gate("Tdagger_3", 3)
-        T_3 = QCOpt.get_full_sized_gate("T_3", 3)
-        cnot_23 = QCOpt.get_full_sized_gate("CNot_2_3", 3)
-        CNot_1_3 = QCOpt.get_full_sized_gate("CNot_1_3", 3)
+        H_3 = QCOpt.get_unitary("H_3", 3)
+        Tdagger_3 = QCOpt.get_unitary("Tdagger_3", 3)
+        T_3 = QCOpt.get_unitary("T_3", 3)
+        cnot_23 = QCOpt.get_unitary("CNot_2_3", 3)
+        CNot_1_3 = QCOpt.get_unitary("CNot_1_3", 3)
 
         return H_3 * cnot_23 * Tdagger_3 * CNot_1_3 * T_3 * cnot_23 * Tdagger_3
     end
 
     return Dict{String, Any}(
-
     "num_qubits" => 3,
     "maximum_depth" => 7,
     "elementary_gates" => ["H_3", "T_1", "T_2", "T_3", "Tdagger_1", "Tdagger_2", "Tdagger_3", "CNot_1_2", "CNot_2_3", "CNot_1_3", "Identity"],
@@ -157,19 +151,18 @@ function toffoli_right()
     # Reference: https://arxiv.org/pdf/0803.2316.pdf 
     
     function target_gate()
-        H_3 = QCOpt.get_full_sized_gate("H_3", 3)
-        Tdagger_2 = QCOpt.get_full_sized_gate("Tdagger_2", 3)
-        T_1 = QCOpt.get_full_sized_gate("T_1", 3)
-        T_2 = QCOpt.get_full_sized_gate("T_2", 3)
-        T_3 = QCOpt.get_full_sized_gate("T_3", 3)
-        CNot_1_2 = QCOpt.get_full_sized_gate("CNot_1_2", 3)
-        CNot_1_3 = QCOpt.get_full_sized_gate("CNot_1_3", 3)
+        H_3 = QCOpt.get_unitary("H_3", 3)
+        Tdagger_2 = QCOpt.get_unitary("Tdagger_2", 3)
+        T_1 = QCOpt.get_unitary("T_1", 3)
+        T_2 = QCOpt.get_unitary("T_2", 3)
+        T_3 = QCOpt.get_unitary("T_3", 3)
+        CNot_1_2 = QCOpt.get_unitary("CNot_1_2", 3)
+        CNot_1_3 = QCOpt.get_unitary("CNot_1_3", 3)
 
         return CNot_1_3 * T_2 * T_3 * CNot_1_2 * H_3 * T_1 * Tdagger_2 * CNot_1_2
     end
 
     return Dict{String, Any}(
-
     "num_qubits" => 3,
     "maximum_depth" => 8,
     "elementary_gates" => ["H_3", "T_1", "T_2", "T_3", "Tdagger_1", "Tdagger_2", "Tdagger_3", "CNot_1_2", "CNot_2_3", "CNot_1_3", "Identity"],
@@ -182,18 +175,17 @@ function miller()
     # Reference: https://doi.org/10.1109/TCAD.2005.858352
 
     function target_gate()
-        CV_1_3 = QCOpt.get_full_sized_gate("CV_1_3", 3)
-        CV_2_3 = QCOpt.get_full_sized_gate("CV_2_3", 3)
-        CVdagger_2_3 = QCOpt.get_full_sized_gate("CVdagger_2_3", 3)
-        CNot_1_2 = QCOpt.get_full_sized_gate("CNot_1_2", 3)
-        CNot_3_1 = QCOpt.get_full_sized_gate("CNot_3_1", 3)
-        CNot_3_2 = QCOpt.get_full_sized_gate("CNot_3_2", 3)
+        CV_1_3 = QCOpt.get_unitary("CV_1_3", 3)
+        CV_2_3 = QCOpt.get_unitary("CV_2_3", 3)
+        CVdagger_2_3 = QCOpt.get_unitary("CVdagger_2_3", 3)
+        CNot_1_2 = QCOpt.get_unitary("CNot_1_2", 3)
+        CNot_3_1 = QCOpt.get_unitary("CNot_3_1", 3)
+        CNot_3_2 = QCOpt.get_unitary("CNot_3_2", 3)
 
         return CNot_3_1 * CNot_3_2 * CV_2_3 * CNot_1_2 * CVdagger_2_3 * CV_1_3 * CNot_3_1 * CNot_1_2
     end
 
     return Dict{String, Any}(
-
     "num_qubits" => 3,
     "maximum_depth" => 8,
     "elementary_gates" => ["CV_1_3", "CV_2_3", "CVdagger_2_3", "CNot_1_2", "CNot_3_1", "CNot_3_2", "Identity"],
@@ -206,7 +198,6 @@ function relative_toffoli()
     #Reference: https://arxiv.org/pdf/1508.03273.pdf
 
     return Dict{String, Any}(
-
         "num_qubits" => 3,
         "maximum_depth" => 9,
         "elementary_gates" => ["H_3", "T_3", "Tdagger_3", "CNot_1_2", "CNot_2_3", "CNot_1_3", "Identity"],
@@ -221,7 +212,6 @@ function margolus()
     #Reference: https://arxiv.org/pdf/quant-ph/0312225.pdf
 
     return Dict{String, Any}(
-
         "num_qubits" => 3,
         "maximum_depth" => 7,
         "elementary_gates" => ["RY_3", "CNot_1_2", "CNot_2_3", "CNot_1_3", "Identity"],
@@ -238,11 +228,27 @@ function CiSwap()
     #Reference: https://doi.org/10.1103/PhysRevResearch.2.033097
 
     return Dict{String, Any}(
-
         "num_qubits" => 3,
         "maximum_depth" => 12,
         "elementary_gates" => ["H_3", "T_3", "Tdagger_3", "CNot_3_2", "CNot_2_3", "CNot_1_3", "Identity"],
         "target_gate" => QCOpt.CiSwapGate(),
+        "objective" => "minimize_depth",
+        "decomposition_type" => "exact_optimal",
+        )
+end
+
+function QFT3()
+
+    # QFT3 circuit using Controlled-Phase gates
+    return Dict{String, Any}(
+        "num_qubits" => 3,
+        "maximum_depth" => 7,   
+        # "elementary_gates" => ["H_1", "H_2", "H_3", "CS_2_1", "CS_3_2", "CT_3_1", "Swap_1_3", "Identity"],
+        "elementary_gates" => ["H_1", "H_2", "H_3", "CU3_2_1", "CU3_3_2", "CU3_3_1", "Swap_1_3", "Identity"],
+        "CU3_θ_discretization" => [0],
+        "CU3_ϕ_discretization" => [0],
+        "CU3_λ_discretization" => [π/2, π/4],
+        "target_gate" => QCOpt.QFT3Gate(),
         "objective" => "minimize_depth",
         "decomposition_type" => "exact_optimal",
         )