+++ examples/optimizers.jl
@@ -3,17 +3,17 @@
 #===================================#
 
 # https://github.com/jump-dev/Gurobi.jl
-function get_gurobi()
+function get_gurobi(; solver_log = false)
     return JuMP.optimizer_with_attributes(Gurobi.Optimizer, 
-                                          MOI.Silent() => false, 
+                                          MOI.Silent() => !solver_log, 
                                         #   "MIPFocus" => 3, # Focus on optimality over feasibility 
                                           "Presolve" => 1) 
 end
 
 # https://github.com/jump-dev/CPLEX.jl
-function get_cplex()
+function get_cplex(; solver_log = true)
      return JuMP.optimizer_with_attributes(CPLEX.Optimizer, 
-                                           MOI.Silent() => false, 
+                                           MOI.Silent() => !solver_log, 
                                            # "CPX_PARAM_EPGAP" => 1E-4,
                                            # "CPX_PARAM_MIPEMPHASIS" => 2 # Focus on optimality over feasibility 
                                            "CPX_PARAM_PREIND" => 1)
@@ -23,33 +23,33 @@ end
 # MIP solvers (open-source, slower)  #
 #====================================#
 # https://github.com/jump-dev/HiGHS.jl
-function get_highs()
+function get_highs(; solver_log = true)
     return JuMP.optimizer_with_attributes(
         HiGHS.Optimizer,
         "presolve" => "on",
-        "log_to_console" => true,
+        "log_to_console" => solver_log,
     )
 end
 
 # https://github.com/jump-dev/Cbc.jl
-function get_cbc()
+function get_cbc(; solver_log = true)
     return JuMP.optimizer_with_attributes(Cbc.Optimizer, 
-                                          MOI.Silent() => false) 
+                                          MOI.Silent() => !solver_log) 
 end
 
 # https://github.com/jump-dev/Glpk.jl
-function get_glpk()
+function get_glpk(; solver_log = true)
     return JuMP.optimizer_with_attributes(GLPK.Optimizer, 
-                                          MOI.Silent() => false)
+                                          MOI.Silent() => !solver_log)
 end
 
 #========================================================#
 # Continuous nonlinear programming solver (open-source)  #
 #========================================================#
 # https://github.com/jump-dev/Ipopt.jl
-function get_ipopt()
+function get_ipopt(; solver_log = true)
      return JuMP.optimizer_with_attributes(Ipopt.Optimizer, 
-                                       MOI.Silent() => true, 
+                                       MOI.Silent() => !solver_log, 
                                        "sb" => "yes", 
                                        "max_iter" => Int(1E4))
 end
@@ -58,9 +58,9 @@ end
 # Local mixed-integer nonlinear programming solver (open-source)  #
 #=================================================================#
 # https://github.com/lanl-ansi/Juniper.jl
-function get_juniper()
+function get_juniper(; solver_log = true)
      return JuMP.optimizer_with_attributes(Juniper.Optimizer, 
-                                         MOI.Silent() => false, 
+                                         MOI.Silent() => !solver_log, 
                                          "mip_solver" => get_gurobi(), 
                                          "nl_solver" => get_ipopt())
 end