+++ src/types.jl
@@ -109,7 +109,7 @@ mutable struct GateData
     "Constructor for struct `GateData`"
     function GateData(gate_type::String, num_qubits::Int64)
         type    = gate_type
-        complex = QCO.get_full_sized_gate(type, num_qubits)
+        complex = QCO.get_unitary(type, num_qubits)
         real    = QCO.complex_to_real_gate(complex)
         inverse = inv(real)
         isreal  = iszero(imag(complex))